Add proc for pattern bindings in match case when needed
Other minor changes and subchanges

Refactor VarPattern to dispatch
Add VPtStruct implementation

TODOs

Discuss if we can divide the cases
where we need pattern bindings and where we don't
just like with Quantity and Entity in the BNFC.
Otherwise, we have this undispatched nested pattern matching.
Alternatively, we can try adding types to the output AST
so we can pattern match the returned (post-dispatch) StrTermCtxt.