Adjust match to generate proc only when a var exists
Implement a hasVariable method for Pattern and ValPattern.
The hasVariable method checks whether a variable exists in a certain Pattern.
For example, the struct `random_struct {2 3 x 4}` contains one variable `x`
and thus it will return true.
The struct `random_struct {1 2 3 4}` contains no variables so it will return false.
More cases might need to be added in the future.