Add comparison operators and var to Quantity in BNFC
Other/Subchanges

Move Var’s under Quantity and Quantity under Entity in BNFC

This allows us to do ring operations on vars that are quantities.
There is no way to differentiate a var that is a quantity
from a var that is an entity (a value with no natural ring operations defined). Update compiler to reflect changes in BNFC without comparison operators.