Update produce/consume to take multiple channels
Since there is no support for tables that do partial key lookups in
Rosette as of now, multiple channels are concatenated into one symbol
and used as a key. A separate lookup table is used to map concatenated
channels to their components and vice versa. I first attempted using a
tuple as a key but that didn’t work as tuples with the same contents do
not hash to the same value if they are not the same actor in Rosette.

A potential area to clean up is to delete items in the lookup table as
they are no longer needed. Currently, new concatenated channels just
get appended to the existing expanded channel list.

Minor adjustments to HelloWorld.rbl and HelloWorldProduceFirst.rbl